Fix critical issues
