+++ shizu/modules/ShizuTester.py
@@ -24,15 +24,12 @@
 
 import time
 import io
-import os
-import sys
-import atexit
 
 import logging
 from .. import logger
 
 from pyrogram import Client, types
-from pyrogram.raw import functions, types as typ
+
 from .. import loader, utils
 
 
@@ -113,59 +110,3 @@ class TesterMod(loader.Module):
         await message.answer(
             self.strings("ping").format(ping),
         )
-
-    async def on_load(self, app: Client):
-        """Create Folder"""
-        if self.db.get("shizu.folder", "folder"):
-            return
-
-        logging.info("Trying to create folder")
-        app.me = await app.get_me()
-        folder_id = 250
-        logs_id = (
-            await utils.create_chat(
-                app,
-                "Shizu-logs",
-                "📫 Shizu-logs do not delete this group, otherwise bot will be broken",
-                True,
-                True,
-                True,
-            )
-        ).id
-
-        backup_id = (
-            await utils.create_chat(
-                app,
-                "Shizu-backup",
-                "📫 Backup-logs do not delete this group, otherwise bot will be broken",
-                True,
-                True,
-                True,
-            )
-        ).id
-
-        logs = await app.resolve_peer(logs_id)
-        backup = await app.resolve_peer(backup_id)
-
-        await app.set_chat_photo(chat_id=logs_id, photo="assets/logs.jpg")
-        await app.set_chat_photo(chat_id=backup_id, photo="assets/backups.jpg")
-
-        await app.invoke(
-            functions.messages.UpdateDialogFilter(
-                id=folder_id,
-                filter=typ.DialogFilter(
-                    id=folder_id,
-                    title="Shizu",
-                    include_peers=[logs, backup],
-                    pinned_peers=[],
-                    exclude_peers=[],
-                    emoticon="❤️",
-                ),
-            )
-        )
-
-        logging.info("Folder created")
-        self.db.set("shizu.folder", "folder", True)
-        self.db.set("shizu.chat", "logs", logs_id)
-        self.db.set("shizu.chat", "backup", backup_id)
-        atexit.register(os.execl(sys.executable, sys.executable, "-m", "shizu"))
